Split into two files
To remove all #[cfg(target_os = "android")] guards from the second file.
